Add copy constructor/assignment and move constructor/assignment test

+++ src/tensor.cpp
@@ -66,7 +66,7 @@ const float* Tensor::data_ptr() const {
 }
 
 float& Tensor::at(std::initializer_list<int> indices) {
-    assert(indices.size() == ndim);
+    assert(indices.size() == static_cast<size_t>(ndim));
     int arr[8];
     int i = 0;
     for (const int& val : indices) {
@@ -74,9 +74,9 @@ float& Tensor::at(std::initializer_list<int> indices) {
     }
     int index = 0;
     int multiple = 1;
-    for (int i = 0; i < indices.size(); i++) {
+    for (size_t i = 0; i < indices.size(); i++) {
         assert(arr[i] < shape[i] && arr[i] >= 0);
-        for (int j = i + 1; j < indices.size(); j++) {
+        for (size_t j = i + 1; j < indices.size(); j++) {
             multiple *= shape[j];
         }
         index += arr[i] * multiple;
